refactor, HV send as percentages

+++ PUTM_EV_Dashboard_2021_soft/Core/Src/main.c
@@ -68,7 +68,7 @@ float lvTempAvg;
 float lvTempMax;
 uint8_t lvState;
 
-float hvVoltage;
+uint8_t hvVoltage;
 float hvTempAvg;
 float hvTempMax;
 uint8_t hvState;
